WIP f-strings

+++ scripts/shared.py
@@ -418,7 +418,7 @@ def get_token_and_span_for_position(editor, position):
     relative_line_number = line_number - logical_line.fFirstLine
     
     token, (x, y) = binary_search.binary_search(
-        tokens, relative_line_number * 10000 + column_number,
+        tokens, int(relative_line_number * 10000 + column_number),
         function=lambda token_and_span:
                          (token_and_span[1][0] * 10000 + token_and_span[1][0]),
         rounding=binary_search.LOW
